NearClip 1.38
Fixes bug where the Nearclip value would be changed twice on world load

Added an option where if you are using the smaller default nearclip (0.001) the mod will raise it to 0.01 when the Quickmenu is open. This can be useful as some worlds case buttons to be hard to press at that value.