Update NearClippingPlaneAdjuster.cs
Added a world blacklist option, the mod will pull a list of worlds at game load and not auto adjust the NearClip plane in those worlds.
Added a toggle for auto adjusting the Near Clip plane. Fixes a bug where setting the NearClip from the QuickMenu when it was 0.001 and RaiseOnQuickMenu was enabled, it would get reset when closing the QM.